+++ ConsoleApp1/Models/Models.cs
@@ -1,11 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TheFiremind.Models
-{
-    internal record Card(string Name, double Cmc, ImageUris Uri, string OracleText, string RulingsUri, string Mana);
-    internal record ImageUris(string Png, string BorderCrop, string Normal);
-}
+﻿namespace TheFiremind.Models;
+record Card(string Name, double Cmc, ImageUris Uri, string OracleText, string RulingsUri, string Mana);
+record ImageUris(string Png, string BorderCrop, string Normal);
+record Ruling(string Source, DateTime Date, string Comment);
+record RulingsList(Ruling[] Rulings);
