--- ConsoleApp1/Models/RulingsList.cs
@@ -1,13 +1,0 @@
-﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TheFiremind.Models;
-
-class RulingsList
-{
-    public Ruling[] Rulings { get; set; } = default!;
-}
