+++ ConsoleApp1/Options/Options.cs
@@ -1,0 +1,3 @@
+﻿namespace TheFiremind.Options;
+
+internal record SettingsOptions(string ScryfallApiBaseUriString, ulong? Ssn1sGuildId);
