--- ConsoleApp1/Options/SettingsOptions.cs
@@ -1,14 +1,0 @@
-﻿namespace TheFiremind.Options;
-
-/// <summary>
-/// 
-/// </summary>
-public class SettingsOptions
-{
-    internal string ScryfallApiBaseUriString { get; set; } = default!;
-    /// <summary>
-    /// 
-    /// </summary>
-    public ulong? Ssn1sGuildId { get; set; }
-}
-}
