+++ ConsoleApp1/Services/CardService.cs
@@ -1,9 +1,4 @@
 ﻿using RestSharp;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using TheFiremind.Models;
 using TheFiremind.Options;
 
@@ -22,7 +17,7 @@ namespace TheFiremind.Services
 
         public Card Get(string name)
         {
-
+            throw new NotImplementedException();
         }
     }
 }
